simple_agent/pyproject.toml
[project]
name = "simpleagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.2",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.1",
    "python-dotenv>=1.1.1",
]
